online_maktaba/requirements/development.txt
-r base.txt
django-debug-toolbar==4.2.0